Backend/requirements.txt
fastapi==0.135.3
groq==1.1.2
passlib==1.7.4
bcrypt
pydantic==2.12.5
pypdf==6.10.0
python-dotenv==1.2.2
python-docx==1.2.0
uvicorn==0.44.0
python-multipart
chromadb
sentence-transformers
pandas

# Security additions
python-jose[cryptography]
slowapi
filetype
